refactor: extract FileContentRenderer component (#507 step 7)
- Move content rendering branches (text/markdown/html/json/jsonl/
  image/pdf/audio/video/binary) into FileContentRenderer.vue
- Move rawUrl() and markdownResult() into the renderer (its only callers)
- Renderer emits markdownLinkClick and updateSource up to FilesView
- Drop now-unused imports (TextResponseView, SchedulerView, TodoExplorer,
  rewriteMarkdownImageRefs, JSON_TOKEN_CLASS, TextResponseData) from FilesView

FilesView.vue now ~343 lines (was 809 pre-refactor).
